Fix mesolve to have a minimal working version (#19)
* Fix duplicated QSolver class

* Add exp_ops argument to mesolve()

* Extract real part for operator expectation value

* Fix exp_save device in odeint

The `get_device()` function returns the device ordinal on GPU and -1
on CPU, raising a runtime error "Device index must not be negative". The
correct way to assign a device is to use the tensor `device` attribute.

* Convert tsave from tensor-like to tensor in mesolve

* Fix exp_ops result ordering

Previous ordering was (i, j) where i is the time index and j the
operator index, now i denotes the operator index and j the time index.
When getting the result from mesolve this is much more natural:
```
states, exp = tq.mesolve(...)
exp[0]  # -> first operator expectation values
```

* Fix fixed step solver time iteration in odeint

The previous version was not saving the latest time result, the saving
code was duplicated, and we couldn't use tqdm. This is all fixed with
this version.

* Add progress bar to fixed time step solver

* Spread utils functions in appropriate files

* Remove unused import

* Fix @gautierronan comments