Minor fixes to `dq.is_herm()` documentation (#515)
* Remove redundant default value documentation

When the type and/or default value is fully specified by the function signature, we don't respecify it in the documentation, to make it lighter. But for example we always document the type of array-like objects to specify their shape.

* Simplify second example for is_herm

More explicit IMO.

* Remove batching-specific documentation

As we do in all other functions, the documentation is written for a non-batched object, and applies *implicitely* to any batch dimension (see e.g. dq.dag, dq.trace, ...).